# openapi.yaml
openapi: "3.0.3"
info:
  title: agent-gateway API
  version: "1.0.0"
  description: |
    Multi-tenant Agent Gateway / Control Plane for LLM agents.
    The only path between your application and the model.
  contact:
    name: agent-gateway team
    email: support@example.com
servers:
  - url: http://localhost:8080/v1
    description: Development server
security:
  - BearerAuth: []
  - SuperAdminAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for tenant users
    SuperAdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for Super Admin
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "unauthorized"
        message:
          type: string
          example: "Invalid or expired token"
        details:
          type: object

    # Auth schemas
    LoginRequest:
      type: object
      required:
        - email
        - password
        - tenant_id
      properties:
        email:
          type: string
          format: email
          example: "admin@acme.com"
        password:
          type: string
          format: password
          example: "strong-password"
        tenant_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
    LoginResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - expires_in
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4..."
        expires_in:
          type: integer
          example: 900
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    SessionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
        user_agent:
          type: string
        ip:
          type: string

    # Tenant schemas
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        status:
          type: string
          enum: [active, suspended, deleted]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # Role schemas
...
paths:
...
  /auth/login:
    post:
      summary: User login
      description: Authenticate with email, password, and tenant_id to get access and refresh tokens
      operationId: login
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Tenant suspended or user inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/super-admin/login:
    post:
      summary: Super Admin login
      description: Authenticate as Super Admin with email and password (no tenant_id required)
      operationId: superAdminLogin
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "admin@platform.com"
                password:
                  type: string
                  format: password
                  example: "strong-password"
      responses:
        '200':
          description: Super Admin login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      summary: Register user in tenant
      description: Register a new user in a tenant (requires tenant_id, password, email)
      operationId: register
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - tenant_id
              properties:
                email:
                  type: string
                  format: email
                  example: "user@acme.com"
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "strong-password"
                tenant_id:
                  type: string
                  format: uuid
                  example: "550e8400-e29b-41d4-a716-446655440000"
                role:
                  type: string
                  example: "operator"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Tenant suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists in tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      summary: Refresh access token
      description: Exchange a refresh token for a new access token (with mandatory rotation)
      operationId: refresh
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Replay detected (token reuse) - entire family revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      summary: Logout (revoke current session)
      description: Revoke the current refresh token / family
      operationId: logout
      tags:
        - auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "logged out"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/sessions:
    get:
      summary: List active sessions
      description: List all active sessions for the current user
      operationId: listSessions
      tags:
        - auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Revoke all sessions
      description: Force revoke all sessions for the current user (compromise recovery)
      operationId: revokeAllSessions
      tags:
        - auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All sessions revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "all sessions revoked"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...